Move card into stable directory and build
