--- src/experimental/components/atoms/CardBody/CardBody.css
@@ -1,3 +1,0 @@
-div {
-  font-family: var(--font-family);
-}
