--- src/experimental/components/atoms/CardBody/cod-card-body.js
@@ -1,2 +1,0 @@
-import CardBody from './CardBody';
-customElements.define('cod-card-body', CardBody);
