--- src/experimental/components/atoms/CardFooter/CardFooter.css
@@ -1,3 +1,0 @@
-div {
-  font-family: var(--font-family);
-}
