--- src/experimental/components/atoms/CardFooter/cod-card-footer.js
@@ -1,2 +1,0 @@
-import CardFooter from './CardFooter';
-customElements.define('cod-card-footer', CardFooter);
