--- src/experimental/components/atoms/CardHeader/CardHeader.css
@@ -1,3 +1,0 @@
-div {
-  font-family: var(--font-family);
-}
