--- src/experimental/components/atoms/CardHeader/cod-card-header.js
@@ -1,2 +1,0 @@
-import CardHeader from './CardHeader';
-customElements.define('cod-card-header', CardHeader);
