--- src/experimental/components/atoms/CardOverlay/CardOverlay.css
@@ -1,3 +1,0 @@
-div {
-  font-family: var(--font-family);
-}
