--- src/experimental/components/atoms/CardOverlay/CardOverlay.js
@@ -1,40 +1,0 @@
-import styles from '!!raw-loader!./CardOverlay.css';
-import varStyles from '!!raw-loader!../../../../shared/variables.css';
-import bootstrapStyles from '!!raw-loader!../../../../shared/themed-bootstrap.css';
-
-const template = document.createElement('template');
-
-template.innerHTML = `
-<slot></slot>
-`;
-
-export default class CardOverlay extends HTMLElement {
-  constructor() {
-    // Always call super first in constructor
-    super();
-    // Create a shadow root
-    const shadow = this.attachShadow({ mode: 'open' });
-    shadow.appendChild(template.content.cloneNode(true));
-    this.cardOverlay = document.createElement('div');
-
-    shadow.addEventListener('slotchange', () => {
-      const tempElements = Array.from(this.children);
-      tempElements.forEach((node) => {
-        this.cardOverlay.append(node);
-      });
-    });
-
-    // Add styles
-    const bootStyles = document.createElement('style');
-    bootStyles.textContent = bootstrapStyles;
-    const variableStyles = document.createElement('style');
-    variableStyles.textContent = varStyles;
-    const itemStyles = document.createElement('style');
-    itemStyles.textContent = styles;
-    shadow.appendChild(bootStyles);
-    shadow.appendChild(variableStyles);
-    shadow.appendChild(itemStyles);
-
-    shadow.appendChild(this.cardOverlay);
-  }
-}
