--- src/experimental/components/atoms/CardOverlay/cod-card-overlay.js
@@ -1,2 +1,0 @@
-import CardOverlay from './CardOverlay';
-customElements.define('cod-card-overlay', CardOverlay);
