+++ src/stable/index-stable.js
@@ -5,3 +5,4 @@ import './components/Drawer/cod-drawer.js';
 import './components/Button/cod-button.js';
 import './components/ButtonGroup/cod-button-group.js';
 import './components/Tag/cod-tag.js';
+import './components/Card/cod-card.js';
